requirements.txt
fastapi
uvicorn[standard]
google-generativeai>=0.5.0
google-genai-client
google-genai
